requirements.txt
django
django-basis
pillow
djangorestframework